feat(codeExplainer): add HTML, CSS, and JavaScript for code explanation feature
- Implemented index.html for the code explainer UI with sections for loading, empty, error, and content states.
- Developed script.js to handle UI interactions, state management, and communication with the VS Code API.
- Created styles.css for styling the code explainer interface, including responsive design and theming based on VS Code variables.
